pyproject.toml
[project]
name = "mlops-labs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

dependencies = [
    "dagshub>=0.5.9",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "hydra-core==1.3.2",
    "joblib>=1.5.0",
    "litserve>=0.2.10",
    "scikit-learn==1.3.2",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pytorch-lightning>=2.5.1.post0",
    "uvicorn>=0.34.2",
]
